refactor: :recycle: http
Refactor http handler
